complete creation paths in api controller
Create the per dag id db components symmetrical to what we create in
terms of dags in the breeder root dag.

We also use the same id to make it straightforward to match to the
individual state from the runtime dags.

We also create a global knowledge state table to consolidate the per dag
state to - named after the overall breeder.