outline optimization archive interplay
We wrap the sampler by querying the archive in case the setting at hand
had already been explored in the past.

Otherwise the previously established exploration process will run.

The dag local sampler will learn the archive stored result also as
though it effectively explored the setting itself in that the archive
stored result is present to its trial as a result.